+++ cookbooks/app-react-context/src/App.tsx
@@ -1,12 +1,21 @@
 import { useModuleCurrentContext } from '@equinor/fusion-framework-react-module-context';
-import { StrictMode } from 'react';
+import { useRelatedContext } from './useRelatedContext';
 
 export const App = () => {
     const { currentContext } = useModuleCurrentContext();
+    // const { value: relatedContext } = useRelatedContext(['EquinorTask']);
+    const { value: relatedContext } = useRelatedContext();
     return (
-        <StrictMode>
-            <pre>{JSON.stringify(currentContext, null, 4)}</pre>
-        </StrictMode>
+        <>
+            <section>
+                <h3>Current Context:</h3>
+                <pre>{JSON.stringify(currentContext, null, 4)}</pre>
+            </section>
+            <section>
+                <h3>Related Context:</h3>
+                <pre>{JSON.stringify(relatedContext, null, 4)}</pre>
+            </section>
+        </>
     );
 };