+++ packages/modules/bookmark/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.2.4 (2023-04-14)
+
+**Note:** Version bump only for package @equinor/fusion-framework-module-bookmark
+
+## 0.2.3 (2023-04-14)
+
+**Note:** Version bump only for package @equinor/fusion-framework-module-bookmark
+
+## 0.2.2 (2023-04-14)
+
+**Note:** Version bump only for package @equinor/fusion-framework-module-bookmark
+
+## [0.2.1](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-module-bookmark@0.2.0...@equinor/fusion-framework-module-bookmark@0.2.1) (2023-04-13)
+
+**Note:** Version bump only for package @equinor/fusion-framework-module-bookmark
+
 ## 0.2.0 (2023-04-11)
 
 ### Features