+++ packages/react/app/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 3.0.25 (2023-04-14)
+
+**Note:** Version bump only for package @equinor/fusion-framework-react-app
+
+## 3.0.24 (2023-04-14)
+
+**Note:** Version bump only for package @equinor/fusion-framework-react-app
+
+## 3.0.23 (2023-04-14)
+
+**Note:** Version bump only for package @equinor/fusion-framework-react-app
+
+## [3.0.22](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-react-app@3.0.21...@equinor/fusion-framework-react-app@3.0.22) (2023-04-13)
+
+**Note:** Version bump only for package @equinor/fusion-framework-react-app
+
 ## 3.0.21 (2023-04-11)
 
 **Note:** Version bump only for package @equinor/fusion-framework-react-app