+++ packages/widget/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.1.16 (2023-04-14)
+
+**Note:** Version bump only for package @equinor/fusion-framework-widget
+
+## 0.1.15 (2023-04-14)
+
+**Note:** Version bump only for package @equinor/fusion-framework-widget
+
+## 0.1.14 (2023-04-14)
+
+**Note:** Version bump only for package @equinor/fusion-framework-widget
+
+## 0.1.13 (2023-04-13)
+
+**Note:** Version bump only for package @equinor/fusion-framework-widget
+
 ## [0.1.12](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-widget@0.1.11...@equinor/fusion-framework-widget@0.1.12) (2023-03-28)
 
 **Note:** Version bump only for package @equinor/fusion-framework-widget